Add files via upload
Cryptography needs to be tested, and there needs to be an addition inserted, I am only as good as I can, criticism if you want to criticize but don't be sentimental, I don't like to be considered as the owner of Syamailcoin, so please fellow Programmers to implement Syamailcoin and I become a Syamailcoin Investor. Don't roll your own crypto.

I admit mistake, I have a strong reason why it needs standalone Cryptography and a Blockchain replacement? When I was in 5th grade where I first fell in love with Math lessons and made a Blueprint or map of life with mathematics, I finally decided to create a new combination formula and it is still relevant in the world of Cryptography or Cryptography keys.. about Heterodyne needs other tools not just code compilation, if it is enough, no problem.. SAI256.. Malicious nodes are denied through Heterodyning or rejected from the criteria structure listed in the Whitepaper... I never said that my latest mathematical formula is the most correct, I never said that SAI256 Cryptography guarantees security without the need to be identified or upgraded or overhauled, in fact I need it and I need you, It Will Pass